+++ R/deprecated.R
@@ -11,10 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-
-text_length <- function(x, filter = NULL, ...)
-{
-    .Deprecated("text_ntoken")
-    text_ntoken(x, filter, ...)
-}